Run programs without extensions
This change gives the ability to run an executable file without an extension. It just replaces your sent bit with a function called shebangtest which will test for a shebang in the first line of the file. If no shebang, then it just runs sent. I also added a few more extensions to the extension list.