i3internet: strip leading whitespace from /proc/net/wireless output
on my system 'grep ^w /proc/net/wireless' doesn't return a result due to
leading space being present in front of wlo1 wifi iface:

"  wlo1: 0000   55.  -55.  -256        0      0      0      0      5"

As a result the signal icon wasn't showing up. Solution is to strip it. Tested on T430s which was ok to begin with, no
regression.